Take the mechanical cast rewrites, and give `MAXLNUM` its real type
Two behaviour-identical passes over the small cast tier, ahead of the
by-hand work that follows.

**`MAXLNUM` is an `i32`.** It was `c_uint = 2147483647` and every one of its
97 use sites cast it back — 72 `as LineNr`, 23 `as c_int`, 1 `as i32`, all
the same 32-bit signed type — plus nine that had already been rewritten to
`MAXLNUM.cast_signed()`. The constant was simply the wrong type: it is one
past the last addressable line, `LineNr` is `i32`, and `i32::MAX` says so.
Every cast at every site is gone. `i32` rather than `LineNr` because
`const_int_alias` counts a `pub const` typed with one of the tree's own
integer aliases, and this would have added a row.

**clippy's machine-applicable suggestions, scoped.** `cast_lossless` and
`ptr_as_ptr` are the only two of the five that carry a `MachineApplicable`
fix (`T::from(x)`, `p.cast::<T>()`), and both are lossless by construction.
Rather than `cargo clippy --fix` tree-wide — 5,113 sites, ~900 files, and no
way to bound what it touches — the suggestions were read out of the
`--force-warn` JSON by byte span and applied only to the 198 files of the
small tier: **223 sites in 99 files**. 27 of them are now at zero.

Five sites needed hand-holding afterwards, all the same trap (p28-5 §4): a
`From::from` is longer than an `as`, so a one-line `unsafe { .. }` reflowed
into five or six and `unsafe_lines` grew. Each was answered by hoisting the
value out of the block rather than by widening it — `startuptime`'s two
`xfree(STARTUPTIME_BUF.replace(..))`, `treesitter`'s `lua_pushnumber`,
`deprecated/eval`'s `array_add`, `gcmd`'s two line reads behind one
`byte_at` closure (which *loses* a line), and `ga_push_string`, which now
borrows its garray once and is two unchecked lines instead of three. Note
`rustfmt`'s `chain_width` (60), not `max_width`: `.cast::<T>().offset(..)`
is a longer *chain* than `(x as *mut T).offset(..)` even when the line is
shorter.

`unsafe_lines_outside_perimeter` 48,358 -> 48,343, `raw_cstr` 5,299 -> 5,290.

+++ crates/nvim/src/diff/getput.rs
@@ -410,7 +410,7 @@ fn diffgetput(
             }
 
             let last = lnum + count - 1 as LineNr;
-            let max = MAXLNUM as c_int as LineNr;
+            let max = MAXLNUM as LineNr;
             let amount = added as LineNr;
             if added != 0 {
                 // SAFETY: the editor exists.
